+++ src/pages/Redirection.jsx
@@ -9,37 +9,9 @@ import { getTodayChallenge } from '../api/challenge';
 const Redirection = () => {
   const navigate = useNavigate();
   const code = new URL(window.location.href).searchParams.get('code');
-  const { setToken, setRefreshToken, setNickname, setLevel } =
+  const { setToken, setRefreshToken, setNickname, setLevelData } =
     useContext(UserContext);
-  const { setChallengeList, setChallengeDate } = useContext(ChallengeContext);
-
-  const setLevelData = async () => {
-    try {
-      const res = await getMember();
-      const data = await res.json();
-
-      setLevel({
-        exp: data.exp,
-        successCnt: data.successCnt,
-        num: data.level.levelNum,
-        name: data.level.name,
-        badge: data.level.badge,
-        expPercent: data.level.expPercent,
-      });
-    } catch (error) {
-      console.error(error);
-    }
-  };
-  const setChallengeData = async () => {
-    try {
-      const res = await getTodayChallenge();
-      const json = await res.json();
-      setChallengeList(json);
-      setChallengeDate(new Date().getDate());
-    } catch (error) {
-      console.error(error);
-    }
-  };
+  const { setChallengeData } = useContext(ChallengeContext);
 
   useEffect(() => {
     (async () => {
@@ -49,20 +21,20 @@ const Redirection = () => {
 
       const json = await res.json();
       localStorage.setItem('kakaoId', json.member.kakaoId);
-      localStorage.setItem('nickname', json.member.nickname);
       localStorage.setItem('memberId', json.member.memberId);
       localStorage.setItem('accessToken', json.token.accessToken);
       localStorage.setItem('refreshToken', json.token.refreshToken);
 
       setToken(json.token.accessToken);
       setRefreshToken(json.token.refreshToken);
-      setNickname(json.member.nickname);
       await setLevelData();
       await setChallengeData();
 
       if (json.member.nickname === null) {
         navigate('/join');
       } else {
+        localStorage.setItem('nickname', json.member.nickname);
+        setNickname(json.member.nickname);
         navigate('/home');
       }
     })();
